Restructures packages, fixes some paths in UserController
