+++ Booze-Blender/src/main/java/boozeblender/BoozeBlenderApplication.java
@@ -1,4 +1,4 @@
-package main.java;
+package boozeblender;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
