+++ Booze-Blender/src/main/java/boozeblender/controllers/AgeVerificationController.java
@@ -1,4 +1,4 @@
-package org.launchcode.controllers;
+package boozeblender.controllers;
 
 public class AgeVerificationController {
 }
