+++ Booze-Blender/src/main/java/boozeblender/controllers/HomepageController.java
@@ -1,4 +1,4 @@
-package org.launchcode.controllers;
+package boozeblender.controllers;
 
 
 public class HomepageController {
