+++ Booze-Blender/src/main/java/boozeblender/controllers/SearchController.java
@@ -1,0 +1,4 @@
+package boozeblender.controllers;
+
+public class SearchController {
+}
