+++ Booze-Blender/src/main/java/boozeblender/controllers/ShoppingListController.java
@@ -1,4 +1,4 @@
-package org.launchcode.controllers;
+package boozeblender.controllers;
 
 public class ShoppingListController {
 }
