+++ Booze-Blender/src/main/java/boozeblender/models/AgeVerification.java
@@ -1,4 +1,4 @@
-package org.launchcode.models;
+package boozeblender.models;
 
 public class AgeVerification {
 }
