+++ Booze-Blender/src/main/java/boozeblender/models/CocktailView.java
@@ -1,4 +1,4 @@
-package org.launchcode.models;
+package boozeblender.models;
 
 public class CocktailView {
 }
