+++ Booze-Blender/src/main/java/boozeblender/models/Header.java
@@ -1,0 +1,4 @@
+package boozeblender.models;
+
+public class Header {
+}
