+++ Booze-Blender/src/main/java/boozeblender/models/Profile.java
@@ -1,0 +1,4 @@
+package boozeblender.models;
+
+public class Profile {
+}
