+++ Booze-Blender/src/main/java/boozeblender/models/Reviews.java
@@ -1,0 +1,4 @@
+package boozeblender.models;
+
+public class Reviews {
+}
