+++ Booze-Blender/src/main/java/boozeblender/models/Search.java
@@ -1,0 +1,4 @@
+package boozeblender.models;
+
+public class Search {
+}
