--- Booze-Blender/src/main/java/org.launchcode/controllers/SearchController.java
@@ -1,4 +1,0 @@
-package org.launchcode.controllers;
-
-public class SearchController {
-}
