--- Booze-Blender/src/main/java/org/launchcode/models/Homepage.java
@@ -1,4 +1,0 @@
-package org.launchcode.models;
-
-public class Homepage {
-}
