--- Booze-Blender/src/main/java/org/launchcode/models/Search.java
@@ -1,4 +1,0 @@
-package org.launchcode.models;
-
-public class Search {
-}
